requirements.txt
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
cssselect==0.9.1
lxml==3.4.2
pymongo==2.8
requests==2.6.0
